cis_user_management/requirements.txt
asgiref==3.8.1
astroid==3.2.4
backports.zoneinfo==0.2.1
dill==0.3.9
Django==4.2.20
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
importlib-metadata==8.5.0
isort==5.13.2
Markdown==3.7
mccabe==0.7.0
platformdirs==4.3.6
PyJWT==2.9.0
pylint==3.2.7
python-decouple==3.8
sqlparse==0.5.3
tomli==2.2.1
tomlkit==0.13.2
typing-extensions==4.13.1
zipp==3.20.2
